Split up onion internal header file
